+++ R/create_analytic_cohort.R
@@ -23,11 +23,10 @@
 #' index_ca_seq = c(1, 2) will return index cancers to
 #' patients with 1 index cancer and will return the first AND second index
 #' cancers to patients with multiple.
-#' @param institution GENIE BPC participating institution. Must be one of
-#' "DFCI", "MSK", "UHN", or "VICC" for NSCLC, BLADDER, Prostate, and PANC cohorts; must be one of "DFCI",
-#' "MSK", "VICC" for CRC and BrCa. Default selection is all institutions.
-#' This parameter corresponds to the variable `institution` in the
-#' Analytic Data Guide.
+#' @param institution GENIE BPC participating institution. See lookup table
+#'  `cohort_institution` for a list of institutions participating in curation of
+#'  each cancer cohort. Default selection is all institutions. This parameter
+#'  corresponds to the variable `institution` in the Analytic Data Guide.
 #' @param stage_dx Stage at diagnosis. Must be one of "Stage I", "Stage II",
 #' "Stage III", "Stage I-III NOS", "Stage IV". The default selection is all
 #' stages.
@@ -155,7 +154,8 @@ create_analytic_cohort <- function(data_synapse,
                                    regimen_order_type,
                                    return_summary = FALSE) {
 
-  # check parameters
+
+# check input parameters --------------------------------------------------
   # cohort object
   if (missing(data_synapse)) {
     stop("Specify the cohort object from the nested list returned by the
@@ -164,30 +164,115 @@ create_analytic_cohort <- function(data_synapse,
     stop("The object specified for data_synapse does not exist.")
   }
 
-  # check input parameter
-  # trying to check that the pull_data_synapse object returned is
-  # specific to the cohort
-  if (!("pt_char" %in% names(data_synapse))) {
-    stop("The data_synapse parameter is expecting a single cohort, e.g., data_synapse_obj$NSCLC_v2.0.
-         Be sure to specify only one cohort at a time, even if there are multiple cohorts
-         in the data_synapse object.")
+  # determine if the user passed a single cancer cohort object
+  # or an object with multiple cohorts pulled
+  # if data_synapse_depth = 4, they supplied the pull_data_synapse() object
+  # if data_synapse_depth = 3, they supplied the data for a single cohort
+  # release, nested within the pull_data_synapse() object
+  # e.g., pull_data_syn_obj$NSCLC_v2.0 (depth 3) vs pull_data_syn_obj (depth 4)
+  data_synapse_depth <- purrr::pluck_depth(data_synapse)
+
+  # if multiple cohorts are supplied, stack the data together before proceeding
+  # 1. convert to data frame (required if one cohort as TM or Rad Tx data and
+  # another does not, need to convert blank list element to dataframe in order
+  # to proceed with the mutate)
+  # 2. Transpose the list to combine each dataset for all cohorts
+  # 3. convert any variables with different types (numeric vs character)
+  # across data releases to be character in order to stack data together
+  # 4. stack corresponding datasets together
+  if (data_synapse_depth == 4){
+    # check input parameter
+    # trying to check that the pull_data_synapse object returned is
+    # specific to the cohort
+    purrr::map(data_synapse, function(x) {
+      names <- names(x)
+
+      if (!("pt_char" %in% names)) {
+        stop(c(
+          "The data_synapse parameter is expecting a nested list of ",
+          "GENIE BPC datasets as returned from `pull_data_synapse()`."
+        ))
+      }
+    })
+
+    # keep only most recent data releases
+    releases_supplied <- tidyr::tibble(cohort_release =
+                                  stringr::str_remove_all(pattern = "-pharma|-consortium|-public",
+                                                 string = names(data_synapse)))
+
+
+    # only keep the most recent release for each cohort
+    releases_supplied_clean <- left_join(releases_supplied,
+                                         synapse_version() %>%
+                                           tidyr::separate(version,
+                                                           into = c("version", "access"),
+                                                           sep = "-") %>%
+                                           dplyr::mutate(cohort_release = paste0(.data$cohort, "_", .data$version)) %>%
+                                           dplyr::select("cohort_release", "release_date"),
+                                         by = c("cohort_release")) %>%
+      tidyr::separate(.data$cohort_release, into = c("cohort", "release"), sep = "_",
+                      remove = FALSE) %>%
+      dplyr::group_by(.data$cohort) %>%
+      dplyr::slice_max(.data$release_date) %>%
+      dplyr::ungroup()
+
+    # check for multiple releases to a single cohort
+    if (length(setdiff(releases_supplied$cohort_release, releases_supplied_clean$cohort_release)) > 0) {
+      cli::cli_inform("Note: Multiple data releases were supplied for the same cancer cohort. Only data corresponding to the most recent data release for each cohort are returned.")
+    }
+
+    # only keep most recent releases before continuing to process
+    data_synapse <- data_synapse %>%
+      purrr::keep(str_remove_all(pattern = "-pharma|-consortium|-public",
+                                 string = names(.))
+                  %in% c(releases_supplied_clean$cohort_release))
+
+    # print message if not all datasets are available for all cohorts
+    dset_datarelease_check <- purrr::map(data_synapse, names) %>%
+      purrr::map_df(., ~tibble(dset = .), .id = "cohort_release") %>%
+      dplyr::group_by(.data$dset) %>%
+      dplyr::summarize(cohorts = paste0(.data$cohort_release, collapse = ", "),
+                       n_cohort_release = n()) %>%
+      # only keep datasets not available for all releases
+      dplyr::filter(.data$n_cohort_release != length(data_synapse))
+
+    if (nrow(dset_datarelease_check) > 0) {
+      cli::cli_inform(paste0("Note: Some datasets are only available for select data releases, including ",
+                             paste0(dset_datarelease_check$dset, " (only available for ", dset_datarelease_check$cohorts, ")", collapse = ", ")))
+    }
+
+    # transpose the list and keep only the most recent data release
+    data_synapse <- purrr::list_transpose(data_synapse) %>%
+      purrr::map_depth(., .depth = 2, as.data.frame) %>%
+      purrr::map_depth(., .depth = 2, ~mutate(.x,
+                                       across(any_of(c("release_version",
+                                                       "naaccr_laterality_cd",
+                                                       "naaccr_tnm_path_desc",
+                                                       "pdl1_iclrange",
+                                                       "pdl1_iclrange_2",
+                                                       "pdl1_icurange",
+                                                       "pdl1_icurange_2",
+                                                       "pdl1_tcurange",
+                                                       "pdl1_lcpsrange",
+                                                       "pdl1_ucpsrange",
+                                                       "cpt_seq_date",
+                                                       "Match_Norm_Seq_Allele1",
+                                                       "Match_Norm_Seq_Allele2",
+                                                       "Protein_position")), ~as.character(.)))) %>%
+      # this only applies when multiple cohorts are supplied
+      purrr::map(., bind_rows, .id = "cohort_release") %>%
+      purrr::map(., mutate, cohort_release = str_remove_all(.data$cohort_release,
+                                                            pattern = "-public|-consortium|-pharma"))
+
   }
-
-  # if (!(stringr::str_to_upper(cohort) %in% c("NSCLC", "CRC", "BRCA"))) {
-  #   stop("Select from available cancer cohorts:
-  #        NSCLC, CRC, BrCa (not case sensitive)")
-  # }
-
-  #  if ( sum(!grepl("^NSCLC$", cohort)>0 , !missing(institution_temp) ,
-  # !grepl(c("^DFCI$|^MSK$|^VICC$|^UHN$"), institution_temp)>0 ) >0  ){
 
   # get cohort name and how it is capitalized in the data_synapse object
   cohort_temp <- pull(
-    pluck(data_synapse, "pt_char") %>%
+    purrr::pluck(data_synapse, "pt_char") %>%
       # remove digits to account for Phase 2 Cohorts
-      mutate(cohort_no_digits = stringr::str_remove_all(pattern = "[:digit:]",
+      dplyr::mutate(cohort_no_digits = stringr::str_remove_all(pattern = "[:digit:]",
                               string = .data$cohort)) %>%
-      distinct(.data$cohort_no_digits),
+      dplyr::distinct(.data$cohort_no_digits),
     "cohort_no_digits"
   )
 
@@ -195,10 +280,10 @@ create_analytic_cohort <- function(data_synapse,
   # how they are stored in variable
   # regimen_drugs
   if (!missing(regimen_drugs)) {
-    regimen_drugs_sorted <- map_chr(
+    regimen_drugs_sorted <- purrr::map_chr(
       strsplit(regimen_drugs, ","), ~
-        toString(str_to_lower(str_sort(
-          (str_trim(.x))
+        toString(stringr::str_to_lower(str_sort(
+          (stringr::str_trim(.x))
         )))
     )
   }
@@ -218,37 +303,34 @@ create_analytic_cohort <- function(data_synapse,
     ))
   }
 
-  # participating institutions by cohort
-  if (sum(
-    !missing(institution),
-    grepl("^NSCLC$|^PANC$|^BLADDER$|^PROSTATE$", stringr::str_to_upper(cohort_temp)) > 0
-  ) > 1) {
-    if (sum(!grepl(
-      c("^DFCI$|^MSK$|^VICC$|^UHN$"),
-      stringr::str_to_upper(institution)
-    ) > 0) > 0) {
-      stop("Select from available participating institutions. For NSCLC/PANC/BLADDER/Prostate, the
-           participating institutions were DFCI, MSK, UHN and VICC.")
-    }
-  } else if (sum(!missing(institution), grepl(
-    "^CRC$|^BRCA$",
-    stringr::str_to_upper(cohort_temp)
-  ) > 0) > 1) {
-    if (sum(!grepl(c("^DFCI$|^MSK$|^VICC$"), stringr::str_to_upper(institution))
-    > 0) > 0) {
-      stop("Select from available participating institutions. For CRC/BrCa, the
-           participating institutions were DFCI, MSK and VICC.")
-    }
-  }
+  # check participating institutions by cohort
+  # if no specified cohorts have the institution available
+  if (!missing(institution)){
+    # collect specified cohort and institution
+    ca_inst_specified <- tibble(
+      cohort = cohort_temp,
+      institution = stringr::str_to_upper(institution)
+    )
 
-  if (missing(institution) & stringr::str_to_upper(cohort_temp) %in%
-      stringr::str_to_upper(c("NSCLC", "PANC", "BLADDER", "PROSTATE"))) {
-    institution_temp <- c("DFCI", "MSK", "UHN", "VICC")
-  } else if (missing(institution) &
-    stringr::str_to_upper(cohort_temp) %in% c("CRC", "BRCA")) {
-    institution_temp <- c("DFCI", "MSK", "VICC")
-  } else {
+    # if any institutions specified that are not available for a cohort
+    chk_inst <- setdiff(ca_inst_specified, genieBPC::cohort_institution) %>%
+      mutate(cohort_institution = paste0(.data$cohort, "-", .data$institution))
+
+    if (nrow(chk_inst) == nrow(ca_inst_specified)){
+      stop("The specified institution is not available for any of the cancer cohorts specified. Review the lookup table `cohort_institution` for the list of available institutions by cohort.")
+    } else if (0 < nrow(chk_inst) & nrow(chk_inst) < nrow(ca_inst_specified)) {
+      # if some of the cohorts don't have the specified institution available
+      cli::cli_inform((paste0("The specified institution did not contribute data for all cancer cohorts provided (",
+                     chk_inst$cohort_institution, ")")))
+
+      institution_temp <- stringr::str_to_upper({{ institution }})
+    } else {
     institution_temp <- stringr::str_to_upper({{ institution }})
+  }
+  } else if (missing(institution)){
+    institution_temp <- genieBPC::cohort_institution %>%
+      dplyr::filter(.data$cohort %in% c(cohort_temp)) %>%
+      dplyr::pull(institution)
   }
 
   # to account for unspecified stage
@@ -272,46 +354,37 @@ create_analytic_cohort <- function(data_synapse,
 
   # to account for unspecified histology
   if (missing(histology)) {
-    if (cohort_temp != "BrCa") {
+    if (max(grepl("BrCa", cohort_temp)) == 0) {
       histology_temp <- pull(pluck(data_synapse, "ca_dx_index") %>%
-        distinct(.data$ca_hist_adeno_squamous), .data$ca_hist_adeno_squamous)
+                               distinct(.data$ca_hist_adeno_squamous), .data$ca_hist_adeno_squamous)
     } else {
-      histology_temp <- pull(
-        pluck(data_synapse, "ca_dx_index") %>%
-          distinct(.data$ca_hist_brca),
-        "ca_hist_brca"
-      )
+      # available histologies
+      avail_histology <- data_synapse$ca_dx_index %>%
+        mutate(histology_aggregated = case_when(
+          .data$cohort == "BrCa" ~ ca_hist_brca,
+          TRUE ~ ca_hist_adeno_squamous
+        )) %>%
+        distinct(.data$histology_aggregated)
+
+      histology_temp <- pull(avail_histology, "histology_aggregated")
     }
   } else {
     histology_temp <- {{ histology }}
   }
 
-
-
   # histology mis-specified
   if (!missing(histology) &&
-    cohort_temp != "BrCa" &&
     sum(!grepl(
-      c("^adenocarcinoma$|^squamous cell$|^sarcoma$|^small cell
-                 carcinoma$|^carcinoma$|^other histologies/mixed tumor$"),
-      stringr::str_to_lower(histology)
-    ) > 0) > 0) {
+      c("^adenocarcinoma$|^squamous cell$|^sarcoma$|^small cell carcinoma$|^carcinoma$|^other histologies/mixed tumor$|^invasive lobular carcinoma$|^invasive ductal carcinoma$|^other histology$"),
+      histology, ignore.case = TRUE) > 0) > 0) {
     stop("Select from available histology categories: Adenocarcinoma,
          Squamous cell, Sarcoma, Small cell carcinoma, Other histologies/mixed
-         tumor")
-  }
-  if (!missing(histology) &&
-    cohort_temp == "BrCa" &&
-    sum(!grepl(
-      c("^invasive lobular carcinoma$|^invasive ductal carcinoma$|
-                 ^Other histology$"),
-      stringr::str_to_lower(histology)
-    ) > 0) > 0) {
-    stop("Select from available histology categories: Invasive lobular
-         carcinoma, Invasive ductal carcinoma, Other histology")
+         tumor (NSCLC, CRC, PANC, Prostate, BLADDER) or Invasive lobular
+         carcinoma, Invasive ductal carcinoma, Other histology (BrCa).")
   }
 
-  ### drug regimen parameter checks
+
+# drug regimen parameter checks -------------------------------------------
   # if regimen type is mis-specified
   if (!missing(regimen_type) | is.numeric(regimen_type)) {
     if (!(stringr::str_to_lower(regimen_type) %in% c("exact", "containing"))) {
@@ -355,16 +428,19 @@ create_analytic_cohort <- function(data_synapse,
     regimen_order_type <- NULL
   }
 
-  ##############################################################################
-  #                             pull cancer cohort                             #
-  ##############################################################################
+
+# pull cancer cohort ------------------------------------------------------
   # select patients based on cohort, institution, stage at diagnosis,
   # histology and cancer number
-  if (cohort_temp != "BrCa") {
+  if (!purrr::is_empty(grep("BrCa", cohort_temp))) {
     cohort_ca_dx <- pluck(data_synapse, "ca_dx_index") %>%
       # re-number index cancer diagnoses
       dplyr::group_by(.data$cohort, .data$record_id) %>%
-      dplyr::mutate(index_ca_seq = 1:n()) %>%
+      dplyr::mutate(index_ca_seq = 1:n(),
+                    histology_aggregated = case_when(
+                      cohort == "BrCa" ~ ca_hist_brca,
+                      TRUE ~ ca_hist_adeno_squamous
+                    )) %>%
       dplyr::ungroup() %>%
       # apply filter(s)
       dplyr::filter(
@@ -372,10 +448,11 @@ create_analytic_cohort <- function(data_synapse,
           stringr::str_to_lower(c(institution_temp)),
         stringr::str_to_lower(.data$stage_dx) %in%
           stringr::str_to_lower(c(stage_dx_temp)),
-        stringr::str_to_lower(.data$ca_hist_adeno_squamous) %in%
+        stringr::str_to_lower(.data$histology_aggregated) %in%
           stringr::str_to_lower(c(histology_temp)),
         .data$index_ca_seq %in% c({{ index_ca_seq }})
-      )
+      ) %>%
+      dplyr::select(-"histology_aggregated")
   } else {
     cohort_ca_dx <- pluck(data_synapse, "ca_dx_index") %>%
       # re-number index cancer diagnoses
@@ -388,7 +465,7 @@ create_analytic_cohort <- function(data_synapse,
           stringr::str_to_lower(c(institution_temp)),
         stringr::str_to_lower(.data$stage_dx) %in%
           stringr::str_to_lower(c(stage_dx_temp)),
-        stringr::str_to_lower(.data$ca_hist_brca) %in%
+        stringr::str_to_lower(.data$ca_hist_adeno_squamous) %in%
           stringr::str_to_lower(c(histology_temp)),
         .data$index_ca_seq %in% c({{ index_ca_seq }})
       )
@@ -396,13 +473,14 @@ create_analytic_cohort <- function(data_synapse,
 
 
 
-  # pull drug regimens to those patients
+
+# pull drug regimens to those patients ------------------------------------
   # option 1: all drug regimens to all patients in cohort
   # regimen_drugs is not specified, regimen_order is not specified
   cohort_ca_drugs <- dplyr::inner_join(cohort_ca_dx %>%
-   dplyr::select("cohort", "record_id", "ca_seq"),
-  pluck(data_synapse, "ca_drugs"),
-  by = c("cohort", "record_id", "ca_seq")
+                                         dplyr::select("cohort", "record_id", "ca_seq"),
+                                       pluck(data_synapse, "ca_drugs"),
+                                       by = c("cohort", "record_id", "ca_seq")
   ) %>%
     # create order for drug regimen within cancer and within times the
     # drug was received
@@ -459,12 +537,12 @@ create_analytic_cohort <- function(data_synapse,
 
     # restrict cancer cohort to all patients who got a drug regimen
     cohort_ca_dx <- dplyr::inner_join(cohort_ca_dx,
-      cohort_ca_drugs %>%
-        dplyr::filter(.data$order_within_cancer %in% c({{ regimen_order }})) %>%
-        dplyr::select("cohort", "record_id", "institution", "ca_seq"),
-      by = c(
-        "cohort", "record_id", "institution", "ca_seq"
-      )
+                                      cohort_ca_drugs %>%
+                                        dplyr::filter(.data$order_within_cancer %in% c({{ regimen_order }})) %>%
+                                        dplyr::select("cohort", "record_id", "institution", "ca_seq"),
+                                      by = c(
+                                        "cohort", "record_id", "institution", "ca_seq"
+                                      )
     )
   }
 
@@ -610,13 +688,13 @@ create_analytic_cohort <- function(data_synapse,
       ) %>%
       # filter on order of interest (e.g. first, all)
       dplyr::filter(.data$order_within_containing_regimen
-        %in% c({{ regimen_order }})) %>%
+                    %in% c({{ regimen_order }})) %>%
       # restrict to patients in the cohort (started with all regimens to all
       # patients)
       dplyr::inner_join(.,
-        cohort_ca_dx %>%
-          dplyr::select("cohort", "record_id", "ca_seq"),
-        by = c("cohort", "record_id", "ca_seq")
+                        cohort_ca_dx %>%
+                          dplyr::select("cohort", "record_id", "ca_seq"),
+                        by = c("cohort", "record_id", "ca_seq")
       ) %>%
       # create blank variables (dropped below, not having them is unique to
       # regimen_order_type = 'containing')
@@ -689,57 +767,58 @@ create_analytic_cohort <- function(data_synapse,
     )
   }
 
-  # for patients meeting the specified criteria, also pull related datasets
+
+# pull other datasets for patients meeting incl criteria ------------------
   # patient characteristics
   cohort_pt_char <- dplyr::inner_join(cohort_ca_dx %>%
-    dplyr::select("cohort", "record_id"),
-  pluck(data_synapse, "pt_char"),
-  by = c("cohort", "record_id")
+                                        dplyr::select("cohort", "record_id"),
+                                      pluck(data_synapse, "pt_char"),
+                                      by = c("cohort", "record_id")
   )
 
   # non-index cancer
   cohort_ca_dx_non_index <- dplyr::inner_join(cohort_ca_dx %>%
-    dplyr::select("cohort", "record_id"),
-  pluck(data_synapse, "ca_dx_non_index"),
-  by = c("cohort", "record_id")
+                                                dplyr::select("cohort", "record_id"),
+                                              pluck(data_synapse, "ca_dx_non_index"),
+                                              by = c("cohort", "record_id")
   )
 
   # PRISSMM Path
   cohort_prissmm_pathology <- dplyr::inner_join(cohort_ca_dx %>%
-    dplyr::select("cohort", "record_id"),
-  pluck(data_synapse, "prissmm_pathology"),
-  by = c("cohort", "record_id")
+                                                  dplyr::select("cohort", "record_id"),
+                                                pluck(data_synapse, "prissmm_pathology"),
+                                                by = c("cohort", "record_id")
   )
 
   # PRISSMM Imaging
   cohort_prissmm_imaging <- dplyr::inner_join(cohort_ca_dx %>%
-    dplyr::select("cohort", "record_id"),
-  pluck(data_synapse, "prissmm_imaging"),
-  by = c("cohort", "record_id")
+                                                dplyr::select("cohort", "record_id"),
+                                              pluck(data_synapse, "prissmm_imaging"),
+                                              by = c("cohort", "record_id")
   )
 
   # PRISSMM Med Onc
   cohort_prissmm_md <- dplyr::inner_join(cohort_ca_dx %>%
-    dplyr::select("cohort", "record_id"),
-  pluck(data_synapse, "prissmm_md"),
-  by = c("cohort", "record_id")
+                                           dplyr::select("cohort", "record_id"),
+                                         pluck(data_synapse, "prissmm_md"),
+                                         by = c("cohort", "record_id")
   )
 
   # TM (if applicable)
   if (!is.null(pluck(data_synapse, "tumor_marker"))) {
     cohort_tumor_marker <- dplyr::inner_join(cohort_ca_dx %>%
-      dplyr::select("cohort", "record_id"),
-    pluck(data_synapse, "tumor_marker"),
-    by = c("cohort", "record_id")
+                                               dplyr::select("cohort", "record_id"),
+                                             pluck(data_synapse, "tumor_marker"),
+                                             by = c("cohort", "record_id")
     )
   }
 
   # RT (if applicable)
   if (!is.null(pluck(data_synapse, "ca_radtx"))) {
     cohort_ca_radtx <- dplyr::inner_join(cohort_ca_dx %>%
-                                               dplyr::select("cohort", "record_id", "ca_seq"),
-                                             pluck(data_synapse, "ca_radtx"),
-                                             by = c("cohort", "record_id", "ca_seq")
+                                           dplyr::select("cohort", "record_id", "ca_seq"),
+                                         pluck(data_synapse, "ca_radtx"),
+                                         by = c("cohort", "record_id", "ca_seq")
     )
   }
 
@@ -780,26 +859,26 @@ create_analytic_cohort <- function(data_synapse,
   # genomic sequencing information
   if (!is.null(pluck(data_synapse, "fusions"))) {
     cohort_fusions <- dplyr::inner_join(pluck(data_synapse, "fusions"),
-      cohort_ngs %>%
-        dplyr::select("cohort", "cpt_genie_sample_id"),
-      by = c("Tumor_Sample_Barcode" = "cpt_genie_sample_id")
+                                        cohort_ngs %>%
+                                          dplyr::select("cohort", "cpt_genie_sample_id"),
+                                        by = c("Tumor_Sample_Barcode" = "cpt_genie_sample_id")
     )
   }
 
   if (!is.null(pluck(data_synapse, "sv"))) {
     cohort_sv <- dplyr::inner_join(pluck(data_synapse, "sv"),
-                                        cohort_ngs %>%
-                                          dplyr::select("cohort", "cpt_genie_sample_id"),
-                                        by = c("Sample_Id" = "cpt_genie_sample_id")
+                                   cohort_ngs %>%
+                                     dplyr::select("cohort", "cpt_genie_sample_id"),
+                                   by = c("Sample_Id" = "cpt_genie_sample_id")
     )
   }
 
   if (!is.null(pluck(data_synapse, "mutations_extended"))) {
     cohort_mutations_extended <- dplyr::inner_join(pluck(data_synapse,
                                                          "mutations_extended"),
-      cohort_ngs %>%
-        dplyr::select("cohort", "cpt_genie_sample_id"),
-      by = c("Tumor_Sample_Barcode" = "cpt_genie_sample_id")
+                                                   cohort_ngs %>%
+                                                     dplyr::select("cohort", "cpt_genie_sample_id"),
+                                                   by = c("Tumor_Sample_Barcode" = "cpt_genie_sample_id")
     )
   }
 
@@ -817,17 +896,25 @@ create_analytic_cohort <- function(data_synapse,
       pull("Tumor_Sample_Barcode")
 
     cohort_cna <- pluck(data_synapse, "cna") %>%
-      select("Hugo_Symbol", any_of(cpt_barcode_keep))
+      select("Hugo_Symbol", any_of(c(cpt_barcode_keep, "cohort_release")))
   }
 
-  # if 0 patients are returned
+  # if 0 patients are returned for all cancer cohorts
   if (nrow(cohort_ca_dx) == 0) {
-    message("No patients meeting the specified criteria were returned.
-            Ensure that all parameters were correctly specified. Specifically,
-            the list of acceptable drugs can be found in the
-            `drug_regimen_list` dataset available with this package.")
+    cli::cli_inform(("No patients meeting the specified criteria were returned.
+            Ensure that all parameters were correctly specified. Review the applicable institution, histology, and regimen drugs parameters to ensure values relevant for all cohorts are supplied. Specifically, the list of acceptable drugs can be found in the `drug_regimen_list` dataset and the list of relevant institutions can be found in the `cohort_institution` dataset available with this package."))
+  } else if (length(setdiff(cohort_temp, cohort_ca_dx$cohort)) > 0){
+    # no patients met criteria for some cancer cohorts
+    cli::cli_inform((paste0("Note: No patients meeting the specified criteria were returned for some cancer cohorts supplied (" ,
+                   paste0(setdiff(cohort_temp, cohort_ca_dx$cohort),
+                          collapse = ", "),
+                   "). \n Review the applicable institution, histology, and regimen drugs parameters to ensure values relevant for all cohorts are supplied. Specifically, the list of acceptable drugs can be found in the `drug_regimen_list` dataset and the list of relevant institutions can be found in the `cohort_institution` dataset available with this package."))
+    )
   }
 
+
+
+# return tables -----------------------------------------------------------
   # return a table 1 to describe the cancer cohort if the user specifies
   if (nrow(cohort_ca_dx) > 0 && return_summary == TRUE) {
 
@@ -841,12 +928,7 @@ create_analytic_cohort <- function(data_synapse,
         data frame",
         type = n_rec_pt ~ "categorical"
       ) %>%
-      gtsummary::modify_header(
-        update = list(
-          stat_0 ~ "**N = {N} patients**"
-        ),
-        quiet = TRUE
-      )
+      gtsummary::modify_header(stat_0 ~ "**N = {N} patients**")
 
     n_rec_drugs_dset <- cohort_ca_drugs %>%
       dplyr::group_by(.data$record_id) %>%
@@ -878,7 +960,7 @@ create_analytic_cohort <- function(data_synapse,
     ) %>%
       gtsummary::bold_labels()
 
-    if (cohort_temp != "BrCa") {
+    if (purrr::is_empty(grep("BrCa", cohort_temp))) {
       tbl_cohort <- cohort_ca_dx %>%
         gtsummary::tbl_summary(
           include = c(
@@ -893,12 +975,7 @@ create_analytic_cohort <- function(data_synapse,
           )
         ) %>%
         gtsummary::bold_labels() %>%
-        gtsummary::modify_header(
-          update = list(
-            stat_0 ~ "**N = {N} Diagnoses**"
-          ),
-          quiet = TRUE
-        )
+        gtsummary::modify_header(stat_0 ~ "**N = {N} Diagnoses**")
     } else {
       tbl_cohort <- cohort_ca_dx %>%
         # dplyr::group_by(.data$record_id) %>%
@@ -915,12 +992,7 @@ create_analytic_cohort <- function(data_synapse,
           )
         ) %>%
         gtsummary::bold_labels() %>%
-        gtsummary::modify_header(
-          update = list(
-            stat_0 ~ "**N = {N} Diagnoses**"
-          ),
-          quiet = TRUE
-        )
+        gtsummary::modify_header(stat_0 ~ "**N = {N} Diagnoses**")
     }
 
     tbl_drugs <- cohort_ca_drugs %>%
@@ -933,12 +1005,7 @@ create_analytic_cohort <- function(data_synapse,
         )
       ) %>%
       gtsummary::bold_labels() %>%
-      gtsummary::modify_header(
-        update = list(
-          stat_0 ~ "**N = {N} Regimens**"
-        ),
-        quiet = TRUE
-      )
+      gtsummary::modify_header(stat_0 ~ "**N = {N} Regimens**")
 
     tbl_ngs <- cohort_ngs %>%
       gtsummary::tbl_summary(
@@ -952,12 +1019,7 @@ create_analytic_cohort <- function(data_synapse,
         )
       ) %>%
       gtsummary::bold_labels() %>%
-      gtsummary::modify_header(
-        update = list(
-          stat_0 ~ "**N = {N} Cancer Panel Tests**"
-        ),
-        quiet = TRUE
-      )
+      gtsummary::modify_header(stat_0 ~ "**N = {N} Cancer Panel Tests**")
   }
 
   # drop variable before returning data frame