+++ R/data.R
@@ -67,6 +67,20 @@
 #' @source \url{https://www.synapse.org/#!Synapse:syn21226493/wiki/599164}
 "synapse_tables"
 
+#' List of Institutions by Cohort
+#'
+#' A dataset containing the institutions participating in curation of each cancer cohort
+#'
+#' @format A lookup table for institutions participating in curation by cancer cohort
+#' \describe{
+#'   \item{cohort}{GENIE BPC Project cancer. One of "NSCLC"
+#'   (non-small cell lung cancer), "CRC" (colorectal cancer),
+#'   "BrCa" (breast cancer), "PANC" (pancreatic cancer),
+#'   "Prostate" (prostate cancer), and "BLADDER" (bladder cancer).}
+#'   \item{institution}{GENIE BPC participating institution.}
+#'   ...
+#' }
+"cohort_institution"
 
 #' Simulated fake GENIE BPC data for function examples and tests
 #'