+++ tests/testthat/test-pull_data_synapse.R
@@ -482,23 +482,23 @@ test_that("Test NA conversion", {
   expect_equal(nrow(any_blank_cols), 0)
 })
 
-# # Consortium: Pull Consortium Data With Username/Password -------------------------------
+# # Consortium: Pull Consortium Data With PAT -------------------------------
 
 test_that("Test class and length of list for public data", {
   skip_if_not(.is_connected_to_genie(pat = Sys.getenv("SYNAPSE_PAT")))
 
   set_synapse_credentials(pat = Sys.getenv("SYNAPSE_PAT"))
 
-  data_releases_username <- synapse_tables %>%
+  data_releases_pat <- synapse_tables %>%
     distinct(cohort, version) %>%
     head(n = 2)
 
-  test_list <- expect_no_error(pmap(select(data_releases_username, cohort, version),
+  test_list <- expect_no_error(pmap(select(data_releases_pat, cohort, version),
                                     pull_data_synapse))
 
 })
 
-# # Consortium: Try Consortium Data With PUBLIC Username/Password -------------------------------
+# # Consortium: Try Consortium Data With PUBLIC PAT -------------------------------
 
 
 test_that("Test that trying to pull consortium with public data fails", {
@@ -587,7 +587,7 @@ test_that("Test class and length of list for public data", {
 
 
 
-# Public: Pull Public Data With Username/Password -----------------------------
+# Public: Pull Public Data With PAT -----------------------------
 test_that("Test no error with Public access PAT, not consortium specific check", {
   skip_if_not(.is_connected_to_genie(pat = Sys.getenv("SYNAPSE_PAT_PUBLIC")))
 
@@ -602,7 +602,7 @@ test_that("Test no error with Public access PAT, not consortium specific check",
 })
 
 
-# No Terms: Pull Public Data With Username/Password -----------------------------
+# No Terms: Pull Public Data With PAT -----------------------------
 
 test_that("Test error when access terms not checked", {
   skip_if_not(.is_connected_to_genie(