+++ NEWS.md
@@ -1,6 +1,10 @@
-# genieBPC (development version)
+# genieBPC 2.1.0
 
-* Update github actions to v4
+* Update `create_analytic_cohort()` to allow cohorts to be created based on multiple cancer types. Previously, only one cancer type at a time was permitted.
+
+* Update GitHub actions to v4
+
+* Breaking change: Remove access to GENIE BPC data on Synapse via username and password per updated Synapse requirements. Access to the data is now via Personal Access Token (PAT) only. 
 
 * Deprecate functions allowing data access by username and password. 
   - Remove `.get_token_by_username()` function